fix: crud frontend

+++ frontEnd_dumbo/src/app/components/dashboard/dashboard.component.ts
@@ -4,6 +4,7 @@ import { ApiUserService } from 'src/app/services/api/api-user.service';
 import { User } from 'src/app/services/auth/user';
 import { Subject } from 'rxjs';
 import { catchError, takeUntil } from 'rxjs/operators';
+import { Router } from '@angular/router';
 
 
 
@@ -17,12 +18,22 @@ export class DashboardComponent implements OnInit{
   rutSearch: string = '';
   private ngUnsubscribe = new Subject();
   dashboardError: string="";
+  filteredUsers: User[] = [];
+  searchForm: FormGroup;
 
-  constructor(private userApiService: ApiUserService, private fb: FormBuilder) {
+  constructor(private userApiService: ApiUserService, private formBuilder: FormBuilder, private router: Router) {
+    this.searchForm = this.formBuilder.group({
+      searchTerm: [''],
+    });
+    this.filteredUsers = [];
   }
 
   ngOnInit(): void {
     this.getUsers();
+  }
+
+  onClickOut(){
+    this.userApiService.logout();
   }
 
   getUsers() {
@@ -34,27 +45,49 @@ export class DashboardComponent implements OnInit{
       })
     ).subscribe((data) => {
       this.users = data;
+      this.filteredUsers = this.users;
+      console.log(this.filteredUsers)
+
     });
+  }
+
+  search() {
+    const searchTerm = this.searchForm.get('searchTerm')?.value.trim().toLowerCase();
+
+    if (searchTerm === '') {
+      // Si el campo de búsqueda está vacío, mostrar todos los usuarios
+      this.filteredUsers = this.users;
+    } else {
+      // Realizar la búsqueda y actualizar los resultados
+      this.userApiService.searchUser(searchTerm).subscribe(
+        (result) => {
+          this.filteredUsers = result;
+        },
+        (error) => {
+          console.error('Error al buscar usuarios:', error);
+        }
+      );
+    }
   }
 
 
 
   deleteUser(userId: number) {
-    this.userApiService.deleteUser(userId).subscribe(
-      () => {
-        this.users = this.users.filter(user => user.id !== userId);
-      },
-      (error) => {
-        console.error('Error al eliminar usuario:', error);
-        // Puedes agregar lógica adicional para manejar el error aquí
-      }
-    );
-  }
+    const confirmDelete = confirm('¿Seguro que quieres eliminar este usuario?');
+    if (confirmDelete) {
+      this.userApiService.deleteUser(userId).subscribe(
+        () => {
+          this.users = this.users.filter(user => user.id !== userId);
+          this.filteredUsers = this.filteredUsers.filter(user => user.id !== userId);
+        },
+        (error) => {
+          console.error('Error al eliminar usuario:', error);
+          // Puedes agregar lógica adicional para manejar el error aquí
+        }
+      );
+    }
 
-  searchUserByRut(rut: string) {
-    // Lógica para buscar usuario por RUT
   }
-
